[REFACTOR] Move support related code to include/tvm/support
- tvm/logging.h -> tvm/support/logging.h
- remove tvm/base.h, move with into tvm/support/with.h
