+++ include/tvm/lowered_func.h
@@ -27,7 +27,6 @@
 
 #include <string>
 
-#include "base.h"
 #include "expr.h"
 #include "tensor.h"
 #include "tvm/node/container.h"
