+++ src/codegen/codegen_cuda.cc
@@ -20,7 +20,7 @@
 /*!
  * \file codegen_cuda.cc
  */
-#include <tvm/base.h>
+
 #include <tvm/runtime/registry.h>
 #include <tvm/packed_func_ext.h>
 #include <cmath>
