+++ src/codegen/llvm/llvm_common.cc
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -22,7 +22,7 @@
  */
 #ifdef TVM_LLVM_VERSION
 
-#include <tvm/base.h>
+#include <dmlc/logging.h>
 #include <atomic>
 #include <mutex>
 #include <memory>
