+++ src/relay/pass/to_a_normal_form.cc
@@ -27,7 +27,7 @@
 #include <tvm/relay/expr_functor.h>
 #include <tvm/relay/transform.h>
 #include <tvm/relay/expr_functor.h>
-#include <tvm/logging.h>
+#include <tvm/support/logging.h>
 #include "let_list.h"
 #include "../../common/arena.h"
 #include "pass_util.h"
